+++ src/pages/dawd/Header/index.css
@@ -1,0 +1,118 @@
+div.header {
+    background: #1c1f31;
+    text-align: left;
+    margin: 0;
+    position: relative;
+    height: 500px;
+    width: 100vw;
+    display: flex;
+    min-height: 500px;
+    color: rgb(183, 183, 183);
+    padding: 0 15%;
+}
+
+img.image {
+    position: absolute;
+    bottom: -15px;
+    left: -80px;
+    max-width: 320px;
+    max-height: 120%;
+    pointer-events: none;
+}
+
+.about-meta {
+    width: 100%;
+    margin: auto;
+    position: relative;
+    padding-left: 10%;
+}
+
+h1#name {
+    font-size: 34px;
+    display: inline;
+    background-color: rgb(183, 183, 183);
+    color: #1c1f31;
+    border-radius: 3px;
+    padding: 0 15px;
+}
+
+h2#title {
+    margin: 0;
+    margin-top: 5px;
+    font-family: 'Nunito';
+    color: rgb(165, 99, 13);
+}
+
+span.separator {
+    margin: 5px 0;
+    height: 2px;
+    display: block;
+    background-color: rgb(165, 99, 13);
+    width: 30%;
+}
+
+p#description {
+    margin: 0;
+    margin-bottom: 30px;
+    max-width: 70%;
+    color: rgb(183, 183, 183);
+    font-family: 'Nunito';
+}
+
+.button-icon {
+    width: 30px;
+    float: left;
+    margin-right: 10px;
+    background: #f0f8ff;
+    display: block;
+    height: 30px;
+    line-height: 30px;
+}
+
+.button-text {
+    margin-left: 5px;
+    line-height: 30px;
+    transition: all 0.3s ease-in-out;
+    font-family: 'Fira Mono';
+}
+
+.header-button {
+    background-color: rgb(15, 15, 15);
+    border-radius: 5px;
+    border: 0;
+    padding: 0;
+    color: rgb(183, 183, 183);
+    width: 180px;
+    text-align: left;
+    cursor: pointer;
+    margin-right: 20px;
+    overflow: hidden;
+    position: relative;
+    font-size: 16px;
+    position: relative;
+    transition: color 0.2s;
+    margin-bottom: 20px;
+
+}
+
+.header-button::before {
+    content: "";
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    width: 0;
+    height: 2px;
+    background-color: transparent;
+    transition: width 0.3s ease-in-out;
+}
+
+.header-button:hover .button-text {
+    color: rgb(165, 99, 13);
+    margin-left: 15px;
+}
+
+.meta-wrapper {
+    z-index: 2;
+    position: relative;
+    text-shadow: -2px 2px 3px black;
+}
